[painter/gl] move #imgToTexture() to common code
It no longer depends on the GL implementation.
